fix errors from PR comments
