+++ src/assets/css/otp-styles-bundle.less
@@ -21,6 +21,7 @@
 @import "./components/cicon";
 @import "../../components/button/cbtn";
 @import "../../components/tabs/ctabs";
+@import "../../components/userpic/cuserpic";
 
 // Responsive utils
 @import "./responsive.less";
