--- src/components/Userpic/stories/Elements/helpers.tsx
@@ -1,16 +1,0 @@
-import * as React from 'react';
-import {Brick} from 'components/utility/Modifiers/Brick';
-
-export const Title: React.FC = ({children}) => {
-    return (
-        <>
-            <span className="t-caption t-caption-16 t-color-gray">{children}</span>
-            <hr className="adivider adivider--thin" />
-            <Brick size={2} />
-        </>
-    );
-};
-
-const imageUrl = 'https://ucarecdn.com/a6b28451-2340-4b61-9776-8e186bf976fa/';
-
-export {Brick, imageUrl};
