--- src/components/Userpic/stories/lib/userpics.less
@@ -1,9 +1,0 @@
-.sb-userpics{
-  background: rgba(255, 255, 255, 0.0001);
-  .content {
-    width: 100%;
-    .separator-bottom {
-      box-shadow: inset 0px -1px 0px rgba(128, 128, 128, 0.2);
-    }
-  }
-}
