+++ src/components/userpic/Pic.tsx
@@ -2,9 +2,9 @@ import * as React from 'react';
 import {MouseEventHandler, ReactNode} from 'react';
 import cx from 'classnames';
 import {component} from '../../services/helpers/classHelpers';
-import './cuserpic.less';
-
 import IconProfile from '../../assets/svg/24/icon-profile-24.svg';
+
+import './index.less';
 
 export interface IProps {
     // image url to use as a background of component
