+++ src/components/userpic/cuserpic.less
@@ -1,5 +1,3 @@
-@import "../../assets/css/bundle.less";
-
 .size(@value) { width: @value; height: @value; }
 
 .cuserpic {
@@ -115,15 +113,15 @@
 
   &&--big {
     .size(48px);
-    font-size: 22.5px;
+    font-size: 22px;
     line-height: 33px;
 
     /* TODO: решение для размеров не является верным.
      * Дизайнеры предполагают, что атомарный объект не имеет своих размеров, на них полностью влияет их контекст
-     * Поэтому пока оставлем так, а новый подход требует обсуждения
+     * Поэтому пока оставляем так, а новый подход требует обсуждения
      */
     @media (max-width: @rsp-mobile-max) {
-      font-size: 20.4px;
+      font-size: 20px;
       line-height: 26px;
 
       width: 40px;
