+++ src/components/userpic/index.less
@@ -1,0 +1,2 @@
+@import "../../assets/css/bundle.less";
+@import "./cuserpic.less";
