--- src/components/utility/Modifiers/Brick.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react';
-import {getSizeForSpacer} from '../../../services/helpers/SpacerHelper';
-
-interface IBrickProps {
-    size?: number;
-    plusHalf?: boolean;
-}
-
-export const Brick: React.FC<IBrickProps> = ({size, plusHalf}: IBrickProps) => {
-    return <div className={getSizeForSpacer('lbrick', size, plusHalf)} />;
-};
