--- src/components/utility/Modifiers/Gap.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react';
-import {getSizeForSpacer} from '../../../services/helpers/SpacerHelper';
-
-interface IGapProps {
-    size?: number;
-    plusHalf?: boolean;
-}
-
-export const Gap: React.FC<IGapProps> = ({size, plusHalf}: IGapProps) => {
-    return <div className={getSizeForSpacer('lgap', size, plusHalf)} />;
-};
