+++ src/index.tsx
@@ -1,4 +1,4 @@
 export * from './components/button';
 export * from './components/tabs';
 export * from './services/helpers/classHelpers';
-export {Orgpic, Userpic} from './components/Userpic';
+export {Orgpic, Userpic} from './components/userpic';
