--- src/services/helpers/SpacerHelper/index.ts
@@ -1,1 +1,0 @@
-export {getSizeForSpacer} from './spacerHelper';
